refactor: KDA-40 이미지 mapper 위치 이동
